fix: text property tab active styles
